# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Duel Standby API
  version: 0.4.0
  description: Auth, dashboard, users, teams, tournaments, treasury, profile, and audit endpoints for DSWeb.
servers:
  - url: https://example.com
components:
  securitySchemes:
    dsAuthCookie:
      type: apiKey
      in: cookie
      name: ds_auth
  schemas:
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - BANNED
    UserRole:
      type: string
      enum:
        - USER
        - MEMBER
        - OFFICER
        - ADMIN
        - FOUNDER
...
    TournamentStatus:
      type: string
      enum:
        - OPEN
        - ONGOING
        - COMPLETED
        - CANCELLED
paths:
...
  /api/users:
    get:
      summary: List users for dashboard management
      security:
        - dsAuthCookie: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - ALL
              - ACTIVE
              - BANNED
        - in: query
          name: role
          schema:
            type: string
            enum:
              - ALL
              - USER
              - MEMBER
              - OFFICER
              - ADMIN
              - FOUNDER
        - in: query
          name: teamId
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: perPage
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list with team filter options
...
  /api/users/{id}/status:
    put:
      summary: Update status, role, or team assignment for a managed user
      security:
        - dsAuthCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/UserStatus"
                role:
                  $ref: "#/components/schemas/UserRole"
                teamId:
                  type:
                    - string
                    - "null"
                reason:
                  type: string
      responses:
        "200":
          description: User updated
        "400":
          description: Validation failed
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /api/treasury:
    get:
      summary: Fetch treasury transactions with summary
      security:
        - dsAuthCookie: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: month
          schema:
            type: integer
        - in: query
          name: year
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
            enum:
              - SEMUA
              - MASUK
              - KELUAR
        - in: query
          name: userId
          schema:
            type: string
      responses:
        "200":
          description: Treasury list
    post:
      summary: Create treasury transaction
      security:
        - dsAuthCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - amount
                - description
              properties:
                type:
                  type: string
                  enum:
                    - MASUK
                    - KELUAR
                amount:
                  type: number
                description:
                  type: string
                userId:
                  type:
                    - string
                    - "null"
      responses:
        "201":
          description: Treasury transaction created
        "403":
          description: Forbidden
  /api/tournaments:
    get:
      summary: List tournaments for public site and dashboard
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/TournamentStatus"
        - in: query
          name: gameType
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Tournament list payload
    post:
      summary: Create tournament
      security:
        - dsAuthCookie: []
      responses:
        "201":
          description: Tournament created
        "403":
          description: Forbidden
  /api/tournaments/{id}:
    get:
      summary: Get tournament detail including participants
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tournament detail payload
        "404":
          description: Not found
    put:
      summary: Update tournament
      security:
        - dsAuthCookie: []
      responses:
        "200":
          description: Tournament updated
        "403":
          description: Forbidden
    delete:
      summary: Delete tournament
      security:
        - dsAuthCookie: []
      responses:
        "200":
          description: Tournament deleted
        "403":
          description: Forbidden
...